+++ core/src/main/java/com/adobe/aem/commons/assetshare/components/details/EditorLinks.java
@@ -19,9 +19,9 @@
 
 package com.adobe.aem.commons.assetshare.components.details;
 
-import org.osgi.annotation.versioning.ProviderType;
+import org.osgi.annotation.versioning.ConsumerType;
 
-@ProviderType
+@ConsumerType
 public interface EditorLinks extends EmptyTextComponent {
     String getAssetFolderEditorPath();