+++ core/src/main/java/com/adobe/aem/commons/assetshare/components/actions/downloads/Downloads.java
@@ -3,11 +3,11 @@ package com.adobe.aem.commons.assetshare.components.actions.downloads;
 import com.adobe.aem.commons.assetshare.content.renditions.download.async.DownloadEntry;
 import com.adobe.cq.dam.download.api.DownloadException;
 import com.adobe.cq.dam.download.api.DownloadProgress;
-import org.osgi.annotation.versioning.ProviderType;
+import org.osgi.annotation.versioning.ConsumerType;
 
 import java.util.List;
 
-@ProviderType
+@ConsumerType
 public interface Downloads {
 
 	/**