+++ core/src/main/java/com/adobe/aem/commons/assetshare/components/details/ActionButtons.java
@@ -19,5 +19,8 @@
 
 package com.adobe.aem.commons.assetshare.components.details;
 
+import org.osgi.annotation.versioning.ConsumerType;
+
+@ConsumerType
 public interface ActionButtons extends EmptyTextComponent {
 }