+++ core/src/main/java/com/adobe/aem/commons/assetshare/components/predicates/PathPredicate.java
@@ -21,11 +21,11 @@ package com.adobe.aem.commons.assetshare.components.predicates;
 
 import com.adobe.cq.wcm.core.components.models.form.OptionItem;
 import com.adobe.cq.wcm.core.components.models.form.Options;
-import org.osgi.annotation.versioning.ProviderType;
+import org.osgi.annotation.versioning.ConsumerType;
 
 import java.util.List;
 
-@ProviderType
+@ConsumerType
 public interface PathPredicate extends Predicate {
 
     /**